Refactor tooltip component.
