+++ step-release-vis/src/app/components/tooltip/tooltip.ts
@@ -1,0 +1,12 @@
+import {Component, OnInit} from '@angular/core';
+
+@Component({
+  selector: 'app-tooltip',
+  templateUrl: './tooltip.html',
+  styleUrls: ['./tooltip.css'],
+})
+export class TooltipComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit(): void {}
+}
